+++ practica-3/ejercicio7c.py
@@ -22,13 +22,13 @@ sobelKernelYAxis = np.array([
 def detect_borders(img, threshold):
     gradientX = cv2.filter2D(img, -1, sobelKernelXAxis)
     gradientY = cv2.filter2D(img, -1, sobelKernelYAxis)
-    magnitude = np.sqrt((gradientX ** 2) + (gradientY ** 2)).astype(np.uint8)
+    magnitude = np.sqrt((gradientX.astype(np.float64) ** 2) + (gradientY.astype(np.float64) ** 2))
     result = magnitude.copy()
 
     result[magnitude <= threshold] = 255
     result[magnitude > threshold] = 0
 
-    return result
+    return result.astype(np.uint8)
 
 def main(argv):
     if (len(argv) < 2):
